Add support for `ON CONFLICT conflict_target DO NOTHING`
This adds support for the next third of #196, adding the ability to
specify the conflict target. This does not yet add support for `DO
UPDATE`, that will come in a later PR. This does not support all
possible conflict targets. They're described in the pg syntax as

```
one of:
    ( { index_column_name | ( index_expression ) } [ COLLATE collation ] [ opclass ] [, ...] ) [ WHERE index_predicate ]
    ON CONSTRAINT constraint_name
```

The grammar supported by this PR is:

```
one of:
    ( index_column_name [, ...] )
    ON CONSTRAINT constraint_name
```

Supporting expression indexes is a bit more work, since column names
have to appear without the table name here. We'd either need to add a
new trait which is implemented for everything in the query builder, or
modify `QueryFragment#to_sql` to have an additional parameter of whether
to include the table name or not.

It's something I want to support eventually, but it's low priority (and
may even have to wait until 2.0). I also need to verify whether bind
parameters can appear in expression position here or not. If the answer
is no, we'll just need to provide a raw SQL API for expression indexes
no matter what.